roomservice: upgrade to the needs of modern era
-- Kanged rooomservice.py from AOKP
-- Adapted for Candy7.1

1. This roomservice supports cm.dependencies and candy.dependencies
2. This roomservice supports changing of dependencies of
   already synced devices
3. This roomservice supports non-CM and non-CANDY dependencies

We assume that the older cm.dependencies format will be supported.
i.e. Not writing 'CyanogenMod/' prefix in repo name is STILL supported.

BUT we also support if the prefix 'CyanogenMod/' 'CandyDevices/' is there.

For non-CANDY and non-CM remotes, you HAVE to write FULL repo name
(roomservice will not modify the repo_name for nonCM, nonCANDY repos).

Change-Id: I31caa17e77f6c6ba58338f4f7ef3000c70a0ad25