vendor: Add masquerade (2/2)
The Masquerade System is the base of what allows
Substratum to run like a Theme Engine. The idea
of a masquerade is to simulate the "server/client"
technicalities and Substratum communicates with the
Masquerade handlers to signal commands directly on
the system.

This will compile masquerade on a global scale to be
built, and shipped, with all devices.

Change-Id: I310ce14ab884ea3ae97e9cb4d0e75a09d817831b