Changes to the new handling of BootAnimations
Update the script to actually calculate and resize each image
to be more appropriate for the device. Before this, the images
would be slightly stretched since it originally hardcoded a ratio
of 1.6 regardless of the devices resolution.

This method does appear to add a couple more seconds to the build
time but if you are going to cry over a few seconds then you should
consider a new hobby outside the Android Community or get a better
build box.

Change-Id: I1fca8f4b9bd2aa4ef2a6825e87de028beabd555e