roomservice: add safeguard for personal local manifests
We support dependency files in the device tree, but there are those
who would prefer to keep using their own local manifest and do not
want the dependency file to interfere with that. Instead of having
them remove their dependency file or "forcing" them to maintain it
we can simply abort reading deps.

 * See if they are using a personal local manifest
 * If any xml file is found and not named, roomservice.xml, and has read permissions then abort
 * This indicates that they do not use their deps file in their device tree
 * If they want to use their deps file, they must remove their current xml file

Change-Id: I8183d246908cd03a3fe7789b089953ee47cc458d